requirements.txt
sentence-transformers
langchain
streamlit
pypdf
python-dotenv
langchain-pinecone
langchain-community
pyairtable
langchain-openai
-e .